[repoView] Enable branch commit command by default
This command adds the abbreviated SHA-1 and message of
the commit at the tip of each branch/tag in the Git Repositories
view.

This was added in 1.0 but was off by default.  It presents
an easy way to see which branches are at the same commits
and so it has enough value to be enabled by default.

Change-Id: Ia1d56476992e5402f4128c477c422d284fae6a20