Select use/create option when projects are in repositories
If all the projects being shared are already in a repository
then preselect the user/create option.

This combined with the project being prechecked in the table
allows the user to share a project without doing any explicit
actions and just clicking 'Finish' button for cases where the
projects they are sharing are already in Git repositories.

Change-Id: If97b765294adb683c7ea29e29da4c1664c0b2508