[historyView] Show commit SHA-1 in first column
Move the Id column from position 3 to position 0 and
show the SHA-1 abbreviated to 7 characters.

Set the minimum width of the first column to show the
entire abbreviated id text.

This should make it easier to give the author/committer
and date columns the remaining space after the message
column but still be able to see the SHA-1 without being
the focus of the content displayed.

Change-Id: If39f8973d3d4b06b8b884bc02e70384e4d56c022