Define target platforms for the build
This simplifies setting up the Eclipse workspace and allows to use
Eclipse versions for development which are not matching the target
platform version. In order to set a target platform open it using the
target editor and click "Set as Target Platform". The EGit Maven build
now also is based on the same target platform definition. This way
dependencies used in the Maven build are identical to those used in
Eclipse.

This change is a prerequisite for change
Id9d721ae574a83ad9440a93ddb143b1271fc2b2b as this is using a feature of
the team framework only available in Juno. Without a target platform
definition this could not be built in older Eclipse versions.

Change-Id: I462e4a35fea4caec7ae95b08ac13cb495c0fa309
JGit-Dependency: Ifc4bfbb3efbab0f5bfbde74f0b2ddc5a2f9ee6a5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>