Correct encoding for compare views of Tree Compare
When the Tree Compare was opened for comparing files from two commits,
the default encoding was used for the individual file compare views.

Now it tries to get a better encoding through
CompareCoreUtils.getResourceEncoding, as we do in other places.

Also, to discourage the use of FileRevisionTypedElement without
specifying an encoding, the constructor without encoding is removed, as
the last usages of it were in Tree Compare.

Bug: 388425
Change-Id: I27c5d715905b65534c24bbced88d95403d4434a0
Signed-off-by: Robin Stocker <robin@nibor.org>