Mark org.eclipse.egit.core bundle to provide project set feature
In order to not prevent upgrade from egit 2.1 when the psf feature was
installed we mark the org.eclipse.egit.core bundle to provision this
feature using p2 instructions.

With this change I was able to upgrade STS 3.1.0 from egit 2.1 to a
local 2.2-SNAPSHOT build which was reported [1] to not work without this
patch.

[1] http://dev.eclipse.org/mhonarc/lists/egit-dev/msg02925.html

Bug: 395031
Change-Id: I0a5f3dc309a79b648ed19fbd37e45e111e36e6d7
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>