Prevent IO_FAILURE by checking for conflicts in ref name validation
E.g. "a/1" is a conflict if "a" already exists, "b" is a conflict if
"b/1" already exists.

With the JGit change that this depends on, we can even show the user
which existing refs cause the conflict.

Bug: 393914
JGit-Dependency: I11e0ba2f300adb8b3612943c304ba68bbe73db8a
Change-Id: Ibea9984121ae88c488858b8a8e73b593195b15e0
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>