Factorize comparison and synchronization code
Use the same code for all comparison actions, making sure that
a comparison launches a synchronization instead when there is more
than a single resource to compare (either when the user selected
more than one resource or when the selected resource is part of a
larger logical model).

This created dead code within CompareUtil and GitSynchronizeModel,
which has been removed or made private.

CQ: 7586
Bug: 416808
Change-Id: I4ae8abfc4dbe7ed772749afa80743278247585ee
Signed-off-by: Laurent Goubet <laurent.goubet@obeo.fr>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>