[stagingView] Fix deletion of files when dragging from Navigator view
In NavigatorDragAdapter, used by the Navigator view, dragFinished
deletes the dragged resources when the operation was DND.DROP_MOVE.

This resulted in the dragged files being deleted when dropping onto the
"Unstaged Changes" or "Staged Changes" in the Staging view.

This was the case even before bug 385412 introduced the possibility to
stage via drag and drop from Package Explorer.

The reason that this bug didn't occur with the Package Explorer is
because it doesn't delete files after DND.DROP_MOVE (which seems like a
more sane behavior).

Bug: 411466
Change-Id: Ibe7891677257d5ff66b7718b728f529134fa5266
Signed-off-by: Robin Stocker <robin@nibor.org>