Offer to create branch when checking out remote-tracking branch
A common mistake for new users of EGit is that they want to work on a
branch existing on a remote and therefore check out the remote-tracking
branch. Instead the right thing to do would have been to create a new
local branch based on it.

This change shows a question whether to create a new local branch or
whether to check out the commit behind the remote-tracking branch.

To make it clear that there will be a question, the label is
"Checkout..." instead of "Checkout" when a remote-tracking branch is
selected.

For advanced users already knowing about this, the question dialog is
not shown when the warning about "detached HEAD" is disabled.

Bug: 341679
Change-Id: I8701aee4bb12de0848f4820a3f7725435cbf106d
Signed-off-by: Robin Stocker <robin@nibor.org>