Fix compare when file is not contained in common ancestor
When a version of a file was compared with another version in three-way
mode and the file was not contained in the common ancestor, the compare
editor did not show the differences but showed the whole files as
conflicting. The user had to switch to two-way compare in order to
see the diffs.

Now the compare editor is always opened in two-way compare mode if there
is no file contained in the common ancestor commit.

Bug: 419134
Change-Id: Ida7210b153880833a71abaf78840eedd69eec3d5
Signed-off-by: Robin Stocker <robin@nibor.org>