Warn when switching branches and a launch is running for the repo
On certain cumbersome platforms (looking at you, Windows), a branch
switch may fail when something is running that uses the files. In case
of launch configurations inside Eclipse, we can detect that and warn the
user about it.

We could even offer to terminate it, but as termination may be async, we
just offer continue/cancel. The user can also choose to not show any
warnings again.

The code for detecting this searches the ISourceContainer of the
launches to see if one is for a project of the repository. This works
e.g. for "Java Application" run configurations and also for servers
(WST). I first tried ILaunchConfiguration#getMappedResources, but it
does not return anything for servers. Maybe there is an easier way, but
I couldn't find it.

See bug for how to test this with servers.

Bug: 419296
Change-Id: I2fd7459b0cf342c13dda087745bd68dc1b6da4c3
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>