Import wizard: Move page init code from getNextPage to setVisible
getNextPage can be called multiple times (and even before the page is
really shown), so getNextPage should not be doing any work for the page.

setVisible seems to be the right place for this, see other wizards. With
this, the logic is only called once and it seems to fix the problem
described in the comments of I23939def6d1a3b88a0812b53fee98a403a664c23.

Change-Id: I5411af8279455921fe9544039d2a94dc8e6f99bf
Signed-off-by: Robin Stocker <robin@nibor.org>