Improve order of menu entries in "Replace With" and "Compare With"
This changes the order by reordering the existing actions. Note that the
order in plugin.xml for the same menubarPath has to be the reverse of
what is desired in the popup menu. It also changes "Replace With > File
in Git Index" to be consistent with "Compare With > Git Index".

The previous changes for this also converted the menu from being defined
via the deprecated org.eclipse.ui.popupMenus extension point to
org.eclipse.ui.menus, but that caused bug 422907 (a bug in Platform UI).

Bug: 421893
Change-Id: I6954f5ad684cc235841e32599e2ef36efa39d0b7
Signed-off-by: Robin Stocker <robin@nibor.org>