Add ability to reword a commit directly from the history view.
This adds a new menu item Modify > Reword to the context menu of
the history view. It can be used to reword a commit, prompting for
a commit message.

An error will be displayed if the selected commit cannot be reached
from the current HEAD commit. This is done so as to avoid the need
to checkout another commit automatically before executing the reword
procedure.

CQ: 8055
Bug: 429365
Change-Id: I51acd62f5c088a539921fd0f297c5e6bded7aef5
Signed-off-by: Maik Schreiber <blizzy@blizzy.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>