Add ability to edit a commit directly from the history view.
This adds a new menu item Modify > Edit to the context menu of the
history view. It can be used to edit a commit, then manually amending
the commit afterwards.

An error will be displayed if the selected commit cannot be reached
from the current HEAD commit. This is done so as to avoid the need
to checkout another commit automatically before executing the edit
procedure.

CQ: 8158
Bug: 431751
Change-Id: Idef404f2d712cccb862a7eb11571a7a4151badb8
Signed-off-by: Maik Schreiber <blizzy@blizzy.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>