Reload GitDocument for Quick Diff in job instead of UI thread
When doing a pull, commit or other operations that change refs, existing
GitDocuments are reloaded. GitDocument is used for Quick Diff, which is
active when the user uses Quick Diff, or when "Show Annotations" was
called.

Doing this possibly expensive operation (it can involve rename
detection) on the UI thread blocks the UI, so do it in a background job
instead. Note that RemoteRevisionQuickDiffProvider (for CVS) of
eclipse.platform.team does it the same way.

Bug: 413497
Change-Id: Iccac6ebce9c63c2ccc95bda28134519d2f379e1b
Signed-off-by: Robin Stocker <robin@nibor.org>