Provide an option to fetch from Gerrit without checking out the branch
Sometimes one wants to fetch a change from gerrit to perform a code
review at a future time.  In that case, it is nice to not checkout the
newly created branch, not only to save time, but also to avoid having to
commit any un-committed changes to the users's current working branch.

By default, this new option will be enabled (perform checkout) to behave
as before.

Change-Id: Ia605a85509db0831e0917a4cab93a519bda3d87f
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>