Add "Stashes" menu to Team context menu and toolbar
Provides a way to stash changes as well as viewing the existing stashes.

The toolbar icon is for stashing changes (when clicking on the icon),
and shows the menu with the existing stashes when using the pull down.

The existing menu item in the repo view was also replaced by this
submenu.

Looks like this when there are no stashes yet:

* Stashes
  * Stash Changes...
  * -
  * (no stashed changes)

And like this with stashes:

* Stashes
  * Stash Changes...
  * -
  * 0: work on foo
  * 1: try fixing bar

Bug: 383794
Bug: 420936
Change-Id: I70a9dba03d569ade9adde49d0ccc09e96f72c237
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>