Refactoring the way how StagingView reacts on selection.
StagingView will ignore all *text* selection changes now. We do not need
to do extra work if user just types in the editor.

Added PartListener which tracks the part activation/deactivation and so
allows to get the selection from the last activated part. As a nice side
effect the view tracks now information if it is visible or not and in
case the view is not visible, *all* selection changes are ignored (but
last one remembered).

The calculation of the repository for selected resource is now done in
background job.

The reload() of the repository is not blocking the caller thread anymore
and is done asynhronously (but still in UI thread).

Bug: 460243
Change-Id: Ie98d3d796e74f00d754e7e4de9716c7bd0990040
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>