Don't suggest branch name for push from remote tracking branch
The "Push Branch..." wizard suggested a wrong target branch name
prefixed with the remote's name if the selected source branch was a
remote tracking branch. Pushing from a remote tracking branch to the
remote branch it's tracking is pointless. So there's no natural
candidate for the target branch. Hence do not suggest any target branch
if pushing a remote tracking branch.

Also improve texts in PushBranchPage to make them easier to understand.

Bug: 464723
Change-Id: I3aed8c30675f1f75efb700fcf568ef91c877d0b5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>