Workaround to enable linking multiple resources to HistoryView
When given a selection the GenericHistoryView only considers the first
element and adapts it to an IResource before passing it on as input. To
work around this behavior the GitHistoryPage now tracks the current
selection itself and uses it as input if preferable.

This does not work for 'Show In->History'; see bug 406894.

Bug: 392948
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Signed-off-by: Axel Richard <axel.richard@obeo.fr>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Alexandra Buzila <abuzila@eclipsesource.com>
Signed-off-by: Philip Langer <planger@eclipsesource.com>
Signed-off-by: Martin Fleck <mfleck@eclipsesource.com>
Change-Id: Idd3f2434881085d90a547f825fb20ba8895b7e79