Fix NPE in GitFlowRepository.isDevelop()
NPE is thrown in isDevelop() if the JGit repository is null.
That repository is only set in the constructor, and I found 
two paths in non-test code, that could lead to this.

To make sure this fixes the issue, I added null guards for both.

Also mark the problematic parameters with null analysis annotations in
order to allow detection of similar problems at compile time

Change-Id: I607ee5c188cf1a949393eea0999bb409c0727f86
Signed-off-by: Max Hohenegger <eclipse@hohenegger.eu>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>