Improve performance of HyperlinkTokenScanner
My initial implementation of this was too inefficient. It invoked the
HyperlinkDetectors for every character, which effectively re-introduced
bug 415537 and made editing long commit messages with only a few links
a pain.

Therefore, compute all hyperlinks on a line and cache them once
computed. Use this cache while we're on that line, and re-compute it
when the current position passes to a new line.

Add an internal subclass of the standard URL hyperlink detector
org.eclipse.jface.text.hyperlink.URLHyperlinkDetector because that one
is only capable of finding hyperlinks at the start of the region.
Others, like those from mylyn, do find all links in a region on their
own.

For an example commit message see: 
https://bugs.eclipse.org/bugs/show_bug.cgi?id=415537

Change-Id: I464f321d5f5d62836d031615b0c6252a54ee893f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>