[nested projects] Allow EGit request inner most projects in some cases
Currently Eclipse resources API offers the choice of having very fast
but unpredictable results (getFileForLocation()) or very slow but
deterministic (findFilesForLocationURI()) while requesting a file via
IPath key in context of nested projects. The "unpredictability" of the
fast API is not documented and has side effects only if workspace
contains nested projects - in this case if the file is located in many
projects, it is not guaranteed, file handle from which project will be
returned (it is related to project names, see bug 476585).

During the fix for bug 440722, by choosing the very fast file lookup for
ResourceUtil we unintentionally changed the EGit behavior so that in
case nested projects were involved, EGit sometimes could not use the
file from the inner most project, but from parent one.

This fix offers an explicit option for clients which would prefer to see
resources from inner most projects independently how long it takes to
resolve them. Some of the affected ResourceUtil API clients are changed
to always request files inner most projects: mostly editor related code.

Bug: 476857
Change-Id: I5f076bbf6bb74392265d6ec1c9c80965111005ef
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>