Fix disappearing scrollbar in RepositorySearchDialog
* Give the directory text input a reasonable minimum width to
  improve the behavior when the dialog is resized ridiculously small.
* Limit the two groups' sizes to at least the sizes of their content.
* Remove the initial miminum height on the tree once it's filled.

Somehow this minimum height setting caused the disappearing scrollbar.
However, if it is removed altogether, the dialog comes up with a tree
that is way too small. Therefore, do set initially a minimum height so
that the dialog as a whole gets a reasonable size, and then remove the
minimum later on.

Bug: 479108
Change-Id: If77a69d6ec83d86a5c84be95de09e478b6ac9f6e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>