Added adapters to IURIEditorInput to better support external files
If we can't get IResource adapter from current editor, IURIEditorInput
is our last resort. It is implemented by FileStoreEditorInput which is
used by Eclipse for editing files which are located outside workspace.

If the URI returned by the IURIEditorInput represents a valid file path,
we can check if we have a repository for this path even if it is not in
the workspace.

Added adapters from IURIEditorInput to java.io.File, Repository and
IHistoryPageSource and updated Repository, Staging and History views so
that they can show their specific content for active external editor.

Change-Id: Ibf582b84b7a8c75662e2ccdfd6453fd1da1876b1
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>